Add a low R variant of secp256k1_ecdsa_sign
Adds a variant of secp256k1_ecdsa_sign which grinds for a signature
with a lower-R. The nonce32 from the previous signature is used as
noncedata for the next signature. With RFC 6979 nonces, the result
is that signatures are still deterministic as the nonce used will be
the one which first results in a valid, lower-R signature.